Move InputComponent to InputGroup
