+++ src/views/pages/authentication/sign-up/SignUpPage.vue
@@ -4,7 +4,7 @@
     import UserInterface from '@/interfaces/UserInterface';
     import validationHandler from '@/core/utilities/validation/validation-hendler';
 
-    const InputComponent = defineAsyncComponent(() => import('@/components/input-components/InputComponent.vue'));
+    const InputGroup = defineAsyncComponent(() => import('@/components/input-components/InputGroup.vue'));
 
     const data = {
         formData: {
@@ -62,14 +62,11 @@
                 <h2 class="mt-5 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">Sign up</h2>
                 <div class="mt-10">
                     <form class="space-y-4" @submit.prevent="handleSubmit">
-                        <div v-for="(input, index) in data.inputs" :key="index">
-                            <InputComponent
-                                v-model="formData[input.inputName]"
-                                :input-name="input.inputName"
-                                :rules="input.rules.join('|')"
-                                :error="validationErrors[input.inputName]"
-                            />
-                        </div>
+                        <InputGroup
+                            :form-data="formData"
+                            :inputs="data.inputs"
+                            :validation-errors="validationErrors"
+                        />
                         <div class="pt-4">
                             <button type="submit" class="flex w-full justify-center rounded-md bg-primary px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Sign up</button>
                         </div>
